pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "promptscope"
version = "0.1.0"
description = "A proof-of-concept library and web app for safe multi-user LLM chats using tool-based context retrieval"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [
    {name = "PromptScope Contributors"}
]
keywords = ["llm", "security", "prompt-injection", "multi-user", "chat"]

classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.2",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
anthropic = [
    "anthropic>=0.39.0",
]
openai = [
    "openai>=1.54.3",
]
all = [
    "anthropic>=0.39.0",
    "openai>=1.54.3",
]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/promptscope"
Repository = "https://github.com/yourusername/promptscope"
Issues = "https://github.com/yourusername/promptscope/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
target-version = "py310"
